Moved nullptr test, and refined docker container details. Also allowed for infinite possible input sizes.
