+++ src/client/test_output.txt
@@ -1,0 +1,5 @@
+0.34404704
+0.3486377
+0.33830002
+0.34094694
+0.33342627
